+++ CHANGELOG.md
@@ -46,7 +46,8 @@ fit: function(response) {
 
 ### v2.6.0 / 2019-01-02
 
-* 接口添加第二个参数，允许动态设置`header`。不用再在`willFetch`中尝试啦！([@alex-mm](https://github.com/alex-mm) in [#69](https://github.com/jias/natty-fetch/issues/69))
+* 接口添加第二个参数，允许动态设置`header`。
+* 修复：如果在`willFetch`中修改了`config`，修改只对本次请求生效。([@alex-mm](https://github.com/alex-mm) in [#69](https://github.com/jias/natty-fetch/issues/69))
 
 ```js
 io.getXxx({